Ranking panel, minor changes, database
Okay so, I did some minor changes to the main GUI files. Changes were
mostly in the functions that did all the visibility for the GUI,
meaning, to make some panels show and others hide. It's nothing too
serious.

I added a reaction panel that shows top 10 scores.

I uploaded the dump of the MySQL database that can be imported, but this
is the tricky part.

The application CANNOT be started without a working MySQL server with
the database imported.
One workaround is to disable the entire rankings panel, because it is
the one that requires a SQL connection.